requirements.txt
pyspark>=3.5.0
mmh3>=4.1.0
kafka-python>=2.0.2
sseclient-py>=1.8.0
requests>=2.31.0
pandas>=2.2.0
matplotlib>=3.8.0
plotly>=5.18.0
streamlit>=1.31.0
pytest>=8.0.0
lxml>=5.1.0
PyYAML>=6.0.2
numpy>=1.26.2
